AI Solution/ai-engine/requirements_ai.txt
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
requests>=2.31.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
geopy>=2.4.0
matplotlib>=3.7.0
seaborn>=0.13.0
ucimlrepo>=0.0.7
xgboost>=2.0.0
